phase3: initial commit, changing project structure
