--- phase1/requirements.txt
@@ -1,1 +1,0 @@
-pytest==3.7.3
