+++ pytest.ini
@@ -1,0 +1,5 @@
+# content of pytest.ini
+
+[pytest]
+testpaths = tests/phase2
+
