+++ tests/phase2/test_login.py
@@ -1,4 +1,3 @@
-from base64 import b64encode
 from http import HTTPStatus
 
 import pytest
